// src/features/editor/enum.ts
export enum FilterEnum {
	Saturation = "saturation",
	Grayscale = "grayscale",
	Polaroid = "polaroid",
	Sepia = "sepia",
	Kodachrome = "kodachrome",
	Contrast = "contrast",
	Brightness = "brightness",
	Brownie = "brownie",
	Vintage = "vintage",
	Technicolor = "technicolor",
	Pixelate = "pixelate",
	Invert = "invert",
	Blur = "blur",
	Sharpen = "sharpen",
	Emboss = "emboss",
	RemoveColor = "removecolor",
	BlackAndWhite = "blacknwhite",
	Vibrance = "vibrance",
	BlendColor = "blendcolor",
	HueRotate = "huerotate",
	Resize = "resize",
	Gamma = "gamma",
}